# api/openapi.yaml
openapi: 3.1.0
info:
  title: Synthesis server
  version: "1.0.0"
  summary: The application service and the relay boundary of the Synthesis server.
  description: |
    The REST surface of `specifications/server/SAS-server-application-service.md`,
    the relay introspection routes of
    `specifications/server/SRB-server-relay-boundary.md`, the project-discovery
    route of `specifications/server/RSN-remote-session.md`, and the three
    WebSocket upgrades of `specifications/server/WSK-websocket.md`.

    Every route and every WebSocket upgrade but `GET /v1/health` needs the header
    `Authorization: Bearer <token>`, carrying `SYNTHESIS_SERVER_TOKEN`. On the
    application routes the token authenticates the persisted administrator user
    (SAS-FR-LFCA, SAS-FR-XRPD), which bypasses every access-control check. On the
    relay it authenticates transport access alone (RSN-FR-TWQJ): the handle
    check, the worker-ownership check, and the end-to-end client authentication
    stand beside it.

    The permission each project-scoped operation needs is named in the
    description of that operation. The relay evaluates no permission.

    The frames the three WebSocket endpoints carry are not described here: a
    WebSocket protocol is outside what OpenAPI describes, and
    `specifications/server/WSK-websocket.md` holds the frame tables.
servers:
  - url: http://localhost:8080
    description: The service as it binds by default.
security:
  - bearerAuth: []
...
paths:
...
  /v1/memberships:
    get:
      tags: [Memberships]
      summary: List memberships, filtered by member or by target.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/MemberIdQuery'
        - $ref: '#/components/parameters/TargetTypeQuery'
        - $ref: '#/components/parameters/TargetIdQuery'
      responses:
        '200':
          description: The window of records.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Membership'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Memberships]
      summary: Create a membership.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMembership'
      responses:
        '201':
          description: The record that was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/invitations:
    get:
      tags: [Invitations]
      summary: List the invitations the caller may read.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: The window of records.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invitation'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Invitations]
      summary: Create an invitation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitation'
      responses:
        '201':
          description: The record that was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/devices:
    get:
      tags: [Devices]
      summary: List the devices the caller owns.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: The window of records.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Devices]
      summary: Register a device.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDevice'
      responses:
        '201':
          description: The record that was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /v1/memberships/{membership_id}:
    parameters:
      - $ref: '#/components/parameters/MembershipId'
    get:
      tags: [Memberships]
      summary: Read one membership.
      responses:
        '200':
          description: The record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Memberships]
      summary: Update one membership.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMembership'
      responses:
        '200':
          description: The record as it now stands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Memberships]
      summary: Delete one membership.
      responses:
        '204':
          description: The record was deleted.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/devices/{device_id}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    get:
      tags: [Devices]
      summary: Read one device.
      responses:
        '200':
          description: The record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Devices]
      summary: Update one device.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDevice'
      responses:
        '200':
          description: The record as it now stands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Devices]
      summary: Delete one device.
      responses:
        '204':
          description: The record was deleted.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /v1/invitations/{invitation_id}:
    parameters:
      - $ref: '#/components/parameters/InvitationId'
    get:
      tags: [Invitations]
      summary: Read one invitation.
      responses:
        '200':
          description: The invitation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/invitations/{invitation_id}/accept:
    parameters:
      - $ref: '#/components/parameters/InvitationId'
    post:
      tags: [Invitations]
      summary: Accept an invitation, once.
      description: The invitation moves to accepted and one active membership is written. An expired, revoked, or accepted invitation is an invitation_not_pending conflict.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInvitation'
      responses:
        '201':
          description: The membership that was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/invitations/{invitation_id}/revoke:
    parameters:
      - $ref: '#/components/parameters/InvitationId'
    post:
      tags: [Invitations]
      summary: Revoke a pending invitation.
      responses:
        '200':
          description: The invitation as it now stands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/devices/{device_id}/revoke:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    post:
      tags: [Devices]
      summary: Revoke a device.
      responses:
        '200':
          description: The device as it now stands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The static token of V1 (SAS-FR-LFCA), read from `SYNTHESIS_SERVER_TOKEN`.
        On the application routes it authenticates the persisted administrator
        user; on the relay routes and the WebSocket upgrades it authenticates
        transport access alone (SRB-FR-JEBM). The service compares it in constant
        time (SAS-FR-BZUK), and it reaches no answer, no log record, and no
        persisted record (SAS-FR-PMRB).
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: The page size. The default is 100 and the maximum is 500.
      schema: { type: integer, minimum: 1, maximum: 500 }
    Offset:
      name: offset
      in: query
      required: false
      description: How many records to pass over before the page starts.
      schema: { type: integer, minimum: 0 }
    MemberIdQuery:
      name: member_id
      in: query
      required: false
      schema: { type: string, format: uuid }
    TargetTypeQuery:
      name: target_type
      in: query
      required: false
      schema: { type: string, enum: [user, team, organization] }
    TargetIdQuery:
      name: target_id
      in: query
      required: false
      schema: { type: string, format: uuid }
...
    MembershipId:
      name: membership_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    InvitationId:
      name: invitation_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    DeviceId:
      name: device_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: |
        A field is absent, empty, or too long; an identifier is not a UUID; or a
        role or a permission is unknown. The codes are invalid_field,
        invalid_role, and invalid_permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthenticated:
      description: The request carries no accepted credential. The code is unauthenticated.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          description: The scheme the route accepts.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The principal may read the record but holds no permission for the operation. The code is forbidden.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        No record holds the identifier, or the principal may not read the
        project, which is answered the same way so the existence of a project is
        not disclosed. The code is not_found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: |
        The write conflicts with the state the service holds. The codes are
        duplicate_id, duplicate_membership, duplicate_grant, owner_protected,
        referenced, stale_revision, conversation_locked, and
        invitation_not_pending.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - unauthenticated
                - forbidden
                - not_found
                - invalid_field
                - invalid_role
                - invalid_permission
                - duplicate_id
                - duplicate_membership
                - duplicate_grant
                - owner_protected
                - referenced
                - stale_revision
                - conversation_locked
                - invitation_not_pending
                - internal
            message: { type: string }
...
    MemberRole:
      type: string
      enum: [owner, administrator, member, viewer]
...
    Status:
      type: string
      enum: [active, suspended]
...
    Membership:
      type: object
      required: [id, member_id, target_type, target_id, role, status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        member_id: { type: string, format: uuid }
        target_type: { type: string, enum: [team, organization] }
        target_id: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/MemberRole' }
        status: { $ref: '#/components/schemas/Status' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateMembership:
      type: object
      required: [member_id, target_type, target_id, role]
      properties:
        id: { type: string, format: uuid }
        member_id: { type: string, format: uuid }
        target_type: { type: string, enum: [team, organization] }
        target_id: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/MemberRole' }
    UpdateMembership:
      type: object
      properties:
        role: { $ref: '#/components/schemas/MemberRole' }
        status: { $ref: '#/components/schemas/Status' }
    Invitation:
      type: object
      required: [id, target_type, target_id, inviter_id, role, expires_at, state, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        target_type: { type: string, enum: [team, organization] }
        target_id: { type: string, format: uuid }
        inviter_id: { type: string, format: uuid }
        invitee_user_id: { type: [string, 'null'], format: uuid }
        invitee_email: { type: [string, 'null'], maxLength: 320 }
        role: { $ref: '#/components/schemas/MemberRole' }
        expires_at: { type: string, format: date-time }
        state: { type: string, enum: [pending, accepted, revoked] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateInvitation:
      type: object
      required: [target_type, target_id, role, expires_at]
      properties:
        id: { type: string, format: uuid }
        target_type: { type: string, enum: [team, organization] }
        target_id: { type: string, format: uuid }
        invitee_user_id: { type: string, format: uuid }
        invitee_email: { type: string, maxLength: 320 }
        role: { $ref: '#/components/schemas/MemberRole' }
        expires_at: { type: string, format: date-time }
    AcceptInvitation:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: The user the membership is written for. It defaults to the invitee.
    Device:
      type: object
      required: [id, owner_id, device_type, display_name, created_at, last_seen_at]
      properties:
        id: { type: string, format: uuid }
        owner_id: { type: string, format: uuid }
        device_type: { type: string, enum: [ide, client] }
        display_name: { type: string, maxLength: 200 }
        created_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        revoked_at: { type: [string, 'null'], format: date-time }
    CreateDevice:
      type: object
      required: [device_type, display_name]
      properties:
        id: { type: string, format: uuid }
        owner_id: { type: string, format: uuid }
        device_type: { type: string, enum: [ide, client] }
        display_name: { type: string, maxLength: 200 }
    UpdateDevice:
      type: object
      properties:
        display_name: { type: string, maxLength: 200 }
        seen: { type: boolean, description: Moves the last-seen time to now. }
...